move source code files into src
