+++ tests/unit/__setup.test.js
@@ -6,8 +6,8 @@ var Sinon = require('sinon');
 
 const PORT = 10000;
 
-var core = require('../../core');
-core.setNativeBindingsModule(path.resolve('native-bindings.js'));
+var core = require('../../src/core');
+core.setNativeBindingsModule(path.resolve('src', 'native-bindings.js'));
 
 global.assert = chai.assert;
 global.TEST_URL = `http://127.0.0.1:${PORT}`;
