+++ tests/unit/timer/setInterval.test.js
@@ -1,5 +1,5 @@
 /* global assert, beforeEach, describe, it */
-const exokit = require('../../../index');
+const exokit = require('../../../src/index');
 
 describe('setInterval', () => {
   var window;
